support simple list
