+++ packages/widget-file-relation/src/preview.tsx
@@ -2,7 +2,10 @@ import * as React from 'react'
 
 export const Preview = ({ value }) => {
   const display = value 
-    ? value.map(item => item.get('label')).join(', ') 
+    ? value.map(item => 
+      (typeof item === 'string') 
+        ? item 
+        : item.get('label')).join(', ') 
     : ''
   return <p>{display}</p>
 }
